EngineManagerに切り出し
